+++ sublet-front/src/@shared/components/Input/TextInputTag.js
@@ -1,9 +1,97 @@
-export const TextInputTag = ({ name = "", value = "", id, label, placeholder, required, onChange }) => (
+import * as s from './Public.styles';
+
+export const InputText = ({ name, placeholder, onChange, value }) => {
+  return (
+    <InputTextCss
+      name={name}
+      type="text"
+      placeholder={placeholder}
+      onChange={onChange}
+      value={value}
+      required
+    />
+  );
+};
+
+export const InputEmail = ({ emailFormatState, onChange, value }) => {
+  return emailFormatState ? (
+    <InputTextCss
+      type="email"
+      name="emailState"
+      placeholder="이메일"
+      onChange={onChange}
+      value={value}
+      required
+    />
+  ) : (
+    <s.InputTextError
+      type="email"
+      name="emailState"
+      placeholder="이메일"
+      onChange={onChange}
+      value={value}
+      required
+    />
+  );
+};
+
+export const InputTelePhone = ({ onChange, value }) => {
+  return (
+    <s.InputText
+      maxlength="13"
+      type="tel"
+      name="phoneState"
+      placeholder="전화번호"
+      onChange={onChange}
+      value={value
+        .replace(/[^0-9]/g, '')
+        .replace(/^(\d{0,3})(\d{0,4})(\d{0,4})$/g, '$1-$2-$3')
+        .replace(/(\-{1,2})$/g, '')}
+      required
+    />
+  );
+};
+
+export const InputPassword = ({ onChange, value }) => {
+  return (
+    <s.InputText
+      type="password"
+      name="passwordState"
+      placeholder="비밀번호"
+      onChange={onChange}
+      value={value}
+      required
+    />
+  );
+};
+
+export const InputStudentId = ({ onChange, value }) => {
+  return (
+    <s.InputText
+      type="tel"
+      maxlength="2"
+      name="studentIdState"
+      placeholder="학번"
+      onChange={inputHandle}
+      value={studentIdState}
+      required
+    />
+  );
+};
+
+export const TextInputTag = ({
+  name = '',
+  value = '',
+  id,
+  label,
+  placeholder,
+  required,
+  onChange,
+}) => (
   <div>
     <label
       htmlFor={id}
-      className="block mb-2 text-sm font-medium text-gray-900 float-left"
-    >
+      className="block mb-2 text-sm font-medium text-gray-900 float-left">
       {label}
     </label>
     {required ? (
@@ -31,12 +119,19 @@ export const TextInputTag = ({ name = "", value = "", id, label, placeholder, re
   </div>
 );
 
-export const TextAreaTag = ({ name = "", value = "", id, label, placeholder, required, handleState }) => (
+export const TextAreaTag = ({
+  name = '',
+  value = '',
+  id,
+  label,
+  placeholder,
+  required,
+  handleState,
+}) => (
   <div>
     <label
       htmlFor={id}
-      className="block mb-2 text-sm font-medium text-gray-900 float-left"
-    >
+      className="block mb-2 text-sm font-medium text-gray-900 float-left">
       {label}
     </label>
     {required ? (
@@ -64,13 +159,19 @@ export const TextAreaTag = ({ name = "", value = "", id, label, placeholder, req
   </div>
 );
 
-
-export const NumberInputTag = ({ id, name = "", label, placeholder, value, required, handleState }) => (
+export const NumberInputTag = ({
+  id,
+  name = '',
+  label,
+  placeholder,
+  value,
+  required,
+  handleState,
+}) => (
   <div>
     <label
       htmlFor={id}
-      className="block mb-2 text-sm font-medium text-gray-900 float-left"
-    >
+      className="block mb-2 text-sm font-medium text-gray-900 float-left">
       {label}
     </label>
     {required ? (
