+++ sublet-front/src/App.js
@@ -20,34 +20,34 @@ import { MobileHeader } from './components/MobileHeader.js';
 import { Desktop, Mobile } from './components/Responsive.js';
 
 function App() {
-	return (
-		<BrowserRouter>
-			<Desktop children={<Header />} />
-			<Mobile children={<MobileHeader />} />
-			<Routes>
-				{/* 기본 페이지 */}
-				<Route path="/" element={<Home />} />
-				<Route path="/roominfo/:roomKey" element={<RoomInfo />} />
-				<Route path="/SearchSubletInfo" element={<SearchSubletInfo />} />
-				<Route path="/ReHome" element={<ReHome />} />
-				<Route path="/Booking" element={<Booking />} />
-				<Route path="/Profile/me" element={<GuestInfo />} />
-				<Route path="/profile/:userId" element={<HostInfo />} />
-				<Route path="/resetpassword" element={<ResetPassword />} />
-				<Route path="/test/chat" element={<TestChat />} />
-				<Route path="/test/chat/:chatId" element={<TestChatRoom />} />
-				<Route path="/test/db/" element={<TestDB />} />
-				<Route path="/test/db/:name" element={<TestDBDetail />} />
-				<Route path="/SaveSublet" element={<SaveSublet />} />
+  return (
+    <BrowserRouter>
+      <Desktop children={<Header />} />
+      <Mobile children={<MobileHeader />} />
+      <Routes>
+        {/* 기본 페이지 */}
+        <Route path="/" element={<Home />} />
+        <Route path="/roominfo/:roomKey" element={<RoomInfo />} />
+        <Route path="/SearchSubletInfo" element={<SearchSubletInfo />} />
+        <Route path="/ReHome" element={<ReHome />} />
+        <Route path="/Booking" element={<Booking />} />
+        <Route path="/Profile/me" element={<GuestInfo />} />
+        <Route path="/profile/:userId" element={<HostInfo />} />
+        <Route path="/resetpassword" element={<ResetPassword />} />
+        <Route path="/test/chat" element={<TestChat />} />
+        <Route path="/test/chat/:chatId" element={<TestChatRoom />} />
+        <Route path="/test/db/" element={<TestDB />} />
+        <Route path="/test/db/:name" element={<TestDBDetail />} />
+        <Route path="/SaveSublet" element={<SaveSublet />} />
 
-				{/* <Route path='/Oauth/kakao' element={<KakaoConfirm />} /> */}
+        {/* <Route path='/Oauth/kakao' element={<KakaoConfirm />} /> */}
 
-				<Route path="/Request" element={<GuestRequest />} />
-				{/* test host page */}
-				<Route path="/CreateSubletInfo" element={<CreateSubletInfo />} />
-			</Routes>
-		</BrowserRouter>
-	);
+        <Route path="/Request" element={<GuestRequest />} />
+        {/* test host page */}
+        <Route path="/CreateSubletInfo" element={<CreateSubletInfo />} />
+      </Routes>
+    </BrowserRouter>
+  );
 }
 
 export default App;
