+++ sublet-front/src/components/Header/Mobile/MobileHeader.js
@@ -1,0 +1,37 @@
+import React from "react";
+import * as s from "./styles/MobileHeader.styles"
+
+export const MobileHeader = () => {
+	const SearchButton = () => {
+		return (
+			<div className="flex items-center">
+				<div className="flex-grow">
+					<s.SearchBut>
+						<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-7 h-7">
+							<path strokeLinecap="round" strokeLinejoin="round" d="m21 21-5.197-5.197m0 0A7.5 7.5 0 1 0 5.196 5.196a7.5 7.5 0 0 0 10.607 10.607Z" />
+						</svg>
+						<div className="flex flex-col">
+							<div className="text-base text-left">어디든지</div>
+							<span className="font-thin">원하는 사람과 날짜</span>
+						</div>
+					</s.SearchBut>
+				</div>
+				<s.CircleIcon>
+					<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
+						<path strokeLinecap="round" strokeLinejoin="round" d="M6 13.5V3.75m0 9.75a1.5 1.5 0 0 1 0 3m0-3a1.5 1.5 0 0 0 0 3m0 3.75V16.5m12-3V3.75m0 9.75a1.5 1.5 0 0 1 0 3m0-3a1.5 1.5 0 0 0 0 3m0 3.75V16.5m-6-9V3.75m0 3.75a1.5 1.5 0 0 1 0 3m0-3a1.5 1.5 0 0 0 0 3m0 9.75V10.5" />
+					</svg>
+				</s.CircleIcon>
+			</div>
+		);
+
+	};
+
+
+
+
+	return (
+		<s.WrapBut>
+			<SearchButton />
+		</s.WrapBut>
+	);
+};
